pyproject.toml
[project]
name = "image-matching-challenge-2025"
version = "0.1.0"
description = "Image Matching Challenge 2025 Solution"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "opencv-python>=4.5.0",
    "matplotlib>=3.4.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "networkx>=2.6.0",
    "tqdm>=4.62.0",
    "tensorboard>=2.6.0",
    "einops>=0.4.0",
    "Pillow>=8.0.0",
]

[project.optional-dependencies]
v2 = [
    "timm>=0.9.0",
    "pytorch-lightning>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src", "src_v2"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
ignore = [
    "E501",   # line too long (let ruff format handle it)
    "PLR0913", # Too many arguments to function call
]